Versione 0.2 di Frienz: + finito di implementare le chiamate al server - utilizza action bar deprecata - da vedere come gestire il socialType e l'id di coupon e campaign - da studiare come capire quando aggiornare il database con nuove immagini di loghi, social e shopping item

+++ projectApp/src/main/java/com/banana/projectapp/profile/SocialAdapter.java
@@ -75,7 +75,8 @@ public class SocialAdapter extends BaseAdapter {
             public void onClick(View v) {
                 Toast.makeText(v.getContext(),"rimuovo "+social.getName(),Toast.LENGTH_SHORT).show();
                 try {
-                    client.deleteSocial((int)social.getId(), DataHolder.getToken());
+                    if (DataHolder.testing)
+                        client.deleteSocial((int)social.getId(), DataHolder.getToken());
                 } catch (EmberTokenInvalid | SocialAccountInvalid | IOException emberTokenInvalid) {
                     emberTokenInvalid.printStackTrace();
                 }
